feat :: make human 폴더 생성 및 Icon 분리 #27
