+++ packages/assets/src/icons/index.ts
@@ -1,34 +1,27 @@
-export * from "./BarChart";
 export * from "./Bell";
 export * from "./Bus";
 export * from "./Calender";
 export * from "./Chart";
 export * from "./Close";
-export * from "./ConvenienceStore";
-export * from "./CookedRice";
-export * from "./CreditCard";
+export * from "./makehuman/CreditCard";
 export * from "./Dev";
 export * from "./DoorOpen";
 export * from "./ForkAndKnife";
-export * from "./FullMoonFace";
 export * from "./Gear";
 export * from "./Home";
-export * from "./LoudSpeaker";
 export * from "./Megaphone";
 export * from "./Menu";
 export * from "./Magnifyingglass";
 export * from "./MoonPlus";
-export * from "./MusicalNote";
 export * from "./Note";
 export * from "./Pen";
 export * from "./Person";
 export * from "./Plus";
-export * from "./SchoolBus";
-export * from "./Tent";
 export * from "./Trash";
 export * from "./arrow";
 export * from "./checkmark";
 export * from "./chevron";
 export * from "./eye";
+export * from "./makehuman";
 export * from "./mark";
 export * from "./notfoundpage";
