+++ packages/assets/src/icons/makehuman/CookedRice.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { StaticIconProps } from "./type";
-import { TENT_PATH_ITEMS } from "./constant"
+import { StaticIconProps } from "../type";
+import { COOKED_RICE_PATH_ITEMS } from "./constant";
 
-export const Tent = ({ size = 24 }: StaticIconProps) => {
+export const CookedRice = ({ size = 24 }: StaticIconProps) => {
   return (
     <svg
       width={size}
@@ -11,7 +11,7 @@ export const Tent = ({ size = 24 }: StaticIconProps) => {
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
-      {TENT_PATH_ITEMS.map((item, idx) => (
+      {COOKED_RICE_PATH_ITEMS.map((item, idx) => (
         <path
           key={idx}
           d={item.d}
