+++ packages/assets/src/icons/makehuman/constant.ts
@@ -1,0 +1,243 @@
+export const BAR_CHART_PATH_ITEMS = [
+  {
+    d: "M1.5 4.5C1.5 2.84315 2.84315 1.5 4.5 1.5H19.5C21.1569 1.5 22.5 2.84315 22.5 4.5V19.5C22.5 20.6104 21.8967 21.5799 21 22.0987L19.9153 21.7498C19.1683 21.5095 18.357 21.5715 17.6551 21.9224L16.5 22.5H14.25L12 21.375L9.75 22.5H7.5L6.34486 21.9224C5.64304 21.5715 4.83167 21.5095 4.08469 21.7498L3 22.0987C2.1033 21.5799 1.5 20.6104 1.5 19.5V4.5Z",
+    fill: "#E1D8EC",
+  },
+  {
+    d: "M7.5 8.25V15.75H1.5V16.5H7.5V22.5H8.25V16.5H15.75V22.5H16.5V16.5H22.5V15.75H16.5V8.25H22.5V7.5H16.5V1.5H15.75V7.5H8.25V1.5H7.5V7.5H1.5V8.25H7.5ZM8.25 8.25H15.75V15.75H8.25V8.25Z",
+    fill: "#B4ACBC",
+  },
+  {
+    d: "M4.5 22.5H7.5V7.6601C7.5 7.15616 7.08953 6.75 6.58024 6.75H3.91976C3.41047 6.75 3 7.15616 3 7.6601V22.0987C3.44126 22.3539 3.95357 22.5 4.5 22.5Z",
+    fill: "#00D26A",
+  },
+  {
+    d: "M13.3302 12.75H10.6698C10.1605 12.75 9.75 13.1562 9.75 13.6603V22.5H14.25V13.6603C14.25 13.1562 13.8395 12.75 13.3302 12.75Z",
+    fill: "#F70A8D",
+  },
+  {
+    d: "M16.5 22.5H19.5C20.0464 22.5 20.5587 22.3539 21 22.0987V5.40674C21 4.90466 20.5895 4.5 20.0802 4.5H17.4198C16.9105 4.5 16.5 4.90466 16.5 5.40674V22.5Z",
+    fill: "#00A6ED",
+  },
+] as const;
+
+export const CONVENIENCE_STORE_PATH_ITEMS = [
+  {
+    d: "M5.9925 7.5075H18.015C18.42 7.5075 18.75 7.1775 18.7575 6.765V2.9925C18.7575 2.58 18.42 2.25 18.015 2.25H5.9925C5.58 2.25 5.25 2.5875 5.25 2.9925V6.765C5.25 7.1775 5.5875 7.5075 5.9925 7.5075Z",
+    fill: "#D3D3D3",
+  },
+  {
+    d: "M2.25 9.00751H21.75V21.0075H2.25V9.00751Z",
+    fill: "#D3D3D3",
+  },
+  {
+    d: "M17.25 3.375V6.39C17.25 6.5925 17.085 6.7575 16.8825 6.7575H16.8675C16.665 6.7575 16.5 6.5925 16.5 6.39V5.265H15V6.39C15 6.5925 14.835 6.7575 14.6325 6.7575H14.6175C14.415 6.7575 14.25 6.5925 14.25 6.39V3.375C14.25 3.1725 14.415 3.0075 14.6175 3.0075H14.6325C14.835 3.0075 15 3.1725 15 3.375V4.5075H16.5V3.375C16.5 3.1725 16.665 3.0075 16.8675 3.0075H16.8825C17.085 3.0075 17.25 3.1725 17.25 3.375ZM9.10499 6.7575C9.36749 6.7575 9.50249 6.6225 9.50249 6.36C9.50249 6.0975 9.35999 6 9.09749 6H8.05499L8.66249 5.3775C8.90249 5.145 9.06749 4.9275 9.17249 4.7175C9.26999 4.515 9.32249 4.305 9.32249 4.08C9.32249 3.7275 9.20999 3.465 8.98499 3.2775C8.75999 3.09 8.42999 3 8.00999 3C7.83749 3 7.65749 3.0225 7.48499 3.06C7.31249 3.0975 7.14749 3.165 6.98249 3.2475C6.89249 3.2925 6.82499 3.36 6.79499 3.435C6.75749 3.51 6.74249 3.54 6.75749 3.615C6.77249 3.69 6.80249 3.7575 6.84749 3.8175C6.89249 3.8775 6.95999 3.915 7.03499 3.9375C7.10999 3.96 7.19999 3.945 7.28999 3.9C7.39499 3.84 7.49249 3.8025 7.58999 3.78C7.68749 3.7575 7.77749 3.7425 7.85999 3.7425C8.19749 3.7425 8.37749 3.9225 8.37749 4.185C8.37749 4.29 8.34749 4.395 8.28749 4.5C8.22749 4.605 8.12999 4.7325 7.99499 4.875L7.06499 5.865C6.98999 5.94 6.93749 6.015 6.89249 6.09C6.84749 6.165 6.83999 6.225 6.83999 6.3225C6.83999 6.465 6.87749 6.57 6.94499 6.6375C7.01249 6.7125 7.12499 6.7425 7.27499 6.7425H9.10499V6.7575ZM12.7575 5.6475C12.7425 5.85 12.57 6 12.3675 6H11.9925V6.3675C11.9925 6.57 11.8425 6.7425 11.64 6.7575C11.4225 6.7725 11.2425 6.6 11.2425 6.3825V6H10.11C10.0425 6 9.98249 5.9775 9.92249 5.9475C9.91499 5.94 9.90749 5.9475 9.89999 5.94C9.89249 5.9325 9.88499 5.925 9.87749 5.9175C9.85499 5.9025 9.83999 5.8875 9.82499 5.865C9.80999 5.8425 9.79499 5.8275 9.78749 5.805C9.77999 5.7825 9.76499 5.7675 9.76499 5.7375C9.75749 5.715 9.74999 5.685 9.74999 5.6625C9.74999 5.6475 9.74249 5.64 9.74249 5.625C9.74249 5.61 9.74999 5.6025 9.74999 5.595C9.74999 5.5725 9.75749 5.5425 9.76499 5.52C9.77249 5.4975 9.77999 5.475 9.79499 5.4525C9.80249 5.445 9.80249 5.43 9.80999 5.4225L11.3175 3.1725C11.325 3.165 11.3325 3.1575 11.34 3.15C11.355 3.1275 11.37 3.1125 11.3925 3.0975C11.415 3.0825 11.43 3.0675 11.4525 3.06C11.475 3.045 11.4975 3.0375 11.52 3.03C11.5425 3.0225 11.565 3.015 11.595 3.015C11.61 3.015 11.6175 3.0075 11.6325 3.0075C11.6475 3.0075 11.655 3.015 11.67 3.015C11.6925 3.015 11.715 3.0225 11.7375 3.03C11.76 3.0375 11.7825 3.045 11.805 3.06C11.8125 3.0675 11.8275 3.0675 11.835 3.075C11.8425 3.0825 11.85 3.09 11.8575 3.0975C11.88 3.1125 11.895 3.1275 11.91 3.15C11.925 3.1725 11.94 3.1875 11.9475 3.21C11.955 3.2325 11.97 3.2475 11.97 3.2775C11.9775 3.3 11.985 3.33 11.985 3.3525C11.985 3.3675 11.9925 3.375 11.9925 3.39V5.265H12.375C12.6 5.25 12.7725 5.43 12.7575 5.6475ZM11.25 5.25V4.6125L10.8225 5.25H11.25Z",
+    fill: "#533566",
+  },
+  {
+    d: "M7.5 7.50751H8.25V9.00751H7.5V7.50751ZM15.75 7.50751H16.5V9.00751H15.75V7.50751Z",
+    fill: "#433B6B",
+  },
+  {
+    d: "M10.5 21.0075H3.75V13.95C3.75 13.7025 3.945 13.5075 4.1925 13.5075H10.0575C10.305 13.5075 10.5 13.7025 10.5 13.95V21.0075Z",
+    fill: "#635994",
+  },
+  {
+    d: "M9.75001 14.67V19.845C9.75001 20.07 9.56251 20.2575 9.33751 20.2575H7.92001C7.69501 20.2575 7.50751 20.07 7.50751 19.845V14.67C7.50751 14.445 7.69501 14.2575 7.92001 14.2575H9.33751C9.56251 14.2575 9.75001 14.445 9.75001 14.67Z",
+    fill: "#83CBFF",
+  },
+  {
+    d: "M6.75001 14.67V19.845C6.75001 20.07 6.56251 20.2575 6.33751 20.2575H4.92001C4.69501 20.2575 4.50751 20.07 4.50751 19.845V14.67C4.50751 14.445 4.69501 14.2575 4.92001 14.2575H6.33751C6.56251 14.2575 6.75001 14.445 6.75001 14.67Z",
+    fill: "#83CBFF",
+  },
+  {
+    d: "M13.1175 18.0075H19.14C19.335 18.0075 19.5 17.8425 19.5075 17.64V15.3675C19.5075 15.165 19.3425 15 19.14 15H13.1175C12.915 15 12.75 15.165 12.75 15.3675V17.64C12.75 17.8425 12.915 18.0075 13.1175 18.0075Z",
+    fill: "#83CBFF",
+  },
+  {
+    d: "M21.93 12.0075H2.0775C1.7625 12.0075 1.5 11.7525 1.5 11.43V9.5775C1.5 9.2625 1.755 9 2.0775 9H21.93C22.245 9 22.5075 9.255 22.5075 9.5775V11.43C22.5 11.7525 22.245 12.0075 21.93 12.0075Z",
+    fill: "#F8312F",
+  },
+] as const;
+
+export const COOKED_RICE_PATH_ITEMS = [
+  {
+    d: "M13.9873 22.5057H9.94485C5.28732 22.5057 1.50732 18.7257 1.50732 14.0682V10.1232H22.4248V14.0682C22.4248 18.7257 18.6448 22.5057 13.9873 22.5057Z",
+    fill: "#F8312F",
+  },
+  {
+    d: "M16.44 5.79568C18.825 5.79568 20.7675 7.7382 20.7675 10.1232C20.7675 12.5082 18.825 14.4507 16.44 14.4507H7.485C5.1 14.4507 3.1575 12.5082 3.1575 10.1232C3.1575 7.7382 5.1 5.79568 7.485 5.79568H16.44ZM16.44 4.13818H7.485C4.1775 4.13818 1.5 6.81568 1.5 10.1232C1.5 13.4307 4.1775 16.1082 7.485 16.1082H16.4325C19.74 16.1082 22.4175 13.4307 22.4175 10.1232C22.425 6.82318 19.74 4.13818 16.44 4.13818Z",
+    fill: "#CA0B4A",
+  },
+  {
+    d: "M3.17552 9.21195L3.16214 9.0855C3.14688 8.94113 3.12304 8.79323 3.09584 8.62583L3.08774 8.5761C3.0216 8.17148 2.93069 7.61528 3.08385 7.07556C3.11186 6.97414 3.1363 6.88289 3.15831 6.80066C3.25838 6.42689 3.30862 6.23926 3.41894 6.12893L3.41803 6.12321C3.52838 5.9817 3.61939 5.8555 3.69284 5.72844C3.81095 5.52332 3.90471 5.28538 4.02204 4.98235L4.0419 4.93096C4.21078 4.4936 4.44308 3.89198 4.89964 3.43542C5.34953 2.98553 5.9401 2.7536 6.36819 2.58548L6.41831 2.56578C6.72733 2.44411 6.96502 2.34858 7.1666 2.22764L7.1719 2.22445L7.17725 2.22136C7.37253 2.10853 7.56578 1.95422 7.8156 1.75199L7.85783 1.71774C8.22345 1.42121 8.73135 1.00932 9.3687 0.841208C9.9804 0.675442 10.6108 0.773836 11.0694 0.845416L11.1257 0.854183C11.4506 0.904583 11.7101 0.943253 11.955 0.943253C12.1999 0.943253 12.4593 0.904583 12.7843 0.854183L12.8405 0.845416C13.2992 0.773836 13.9295 0.675441 14.5412 0.841208C15.1786 1.00932 15.6865 1.42121 16.0521 1.71774L16.0943 1.75199C16.344 1.95406 16.5371 2.1083 16.7322 2.2211C16.9373 2.33922 17.1753 2.43299 17.4784 2.55033L17.5297 2.57019C17.9671 2.73907 18.5687 2.97137 19.0253 3.42793C19.4804 3.88306 19.7126 4.48498 19.8806 4.92049L19.9005 4.97201C20.019 5.27829 20.1127 5.51593 20.2306 5.71829L20.2318 5.72054C20.3195 5.87221 20.4322 6.02264 20.5729 6.19931L20.5723 6.20217C20.7663 6.48978 20.7724 6.51401 20.8391 6.77668C20.8581 6.85115 20.8819 6.94477 20.9162 7.06899C21.0693 7.60868 20.9784 8.16488 20.9123 8.5695L20.9042 8.61915C20.8766 8.78888 20.8525 8.9385 20.8372 9.08453L20.825 9.20205C20.7246 11.5879 18.6599 13.5 16.1372 13.5H7.86285C5.34374 13.5 3.28122 11.5931 3.17552 9.21195Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M20.858 9.55402C20.8287 9.74288 20.7992 9.93278 20.7804 10.1221H20.7676V10.1231C20.7676 12.5081 18.8251 14.4506 16.4401 14.4506H7.48509C5.10009 14.4506 3.15759 12.5081 3.15759 10.1231V10.1221H3.14393C3.12506 9.93525 3.09601 9.7479 3.06711 9.56153C2.9896 9.06165 2.91321 8.56905 3.03754 8.1132C3.16346 7.6329 3.48656 7.23373 3.80672 6.83823C4.00143 6.59769 4.19504 6.35851 4.34254 6.10324C4.49453 5.8394 4.60814 5.54597 4.72152 5.25312C4.90468 4.78007 5.08724 4.30853 5.43004 3.96574C5.76779 3.62799 6.23052 3.44578 6.69317 3.26361C6.99076 3.14643 7.28831 3.02927 7.5525 2.87074C7.8078 2.72324 8.04698 2.52962 8.2875 2.33492C8.68305 2.01476 9.0822 1.69166 9.5625 1.56574C10.0184 1.44141 10.511 1.5178 11.0108 1.59531C11.3252 1.64405 11.6423 1.69324 11.955 1.69324C12.2678 1.69324 12.5849 1.64405 12.8993 1.59531C13.3991 1.5178 13.8917 1.44141 14.3475 1.56574C14.8279 1.69166 15.227 2.01476 15.6226 2.33492C15.8631 2.52962 16.1023 2.72324 16.3575 2.87074C16.6214 3.02273 16.9148 3.13634 17.2076 3.24972C17.6807 3.43288 18.1523 3.61544 18.495 3.95824C18.8363 4.29949 19.0188 4.7712 19.2011 5.24258C19.3153 5.5378 19.4294 5.83289 19.5825 6.09574C19.73 6.35101 19.9237 6.59019 20.1184 6.83073C20.4385 7.22623 20.7617 7.6254 20.8875 8.1057C21.0119 8.56155 20.9355 9.05415 20.858 9.55402Z",
+    fill: "#F4F4F4",
+  },
+  {
+    d: "M11.2875 3.35819L12 4.58819C12.1575 4.86569 12.0675 5.21069 11.79 5.36819C11.5125 5.52569 11.1675 5.42819 11.01 5.15819L10.2975 3.92819C10.14 3.65069 10.2375 3.30569 10.5075 3.14819C10.785 2.99069 11.13 3.08819 11.2875 3.35819Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M14.55 5.94577L14.8125 4.55077C14.8725 4.24327 15.1725 4.04077 15.48 4.09327C15.795 4.15327 15.9975 4.45327 15.9375 4.76077L15.675 6.15577C15.615 6.47077 15.315 6.67327 15.0075 6.61327C14.6925 6.55327 14.49 6.25327 14.55 5.94577Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M9.69001 9.4782L9.11251 8.1807C8.98501 7.8882 9.11251 7.5582 9.39751 7.43071C9.68251 7.30321 10.02 7.43071 10.1475 7.7157L10.725 9.0132C10.8525 9.2982 10.725 9.6357 10.44 9.7632C10.155 9.8907 9.81751 9.7632 9.69001 9.4782Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M6.92245 12.7483L7.18495 11.3533C7.23745 11.0383 7.53747 10.8358 7.85247 10.8958C8.16747 10.9558 8.36997 11.2558 8.30997 11.5633L8.04747 12.9583C7.98747 13.2733 7.68747 13.4758 7.37995 13.4158C7.06495 13.3558 6.86245 13.0558 6.92245 12.7483Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M5.86498 8.59321L4.60498 9.25321C4.32748 9.39571 4.21498 9.74071 4.36498 10.0182C4.50748 10.2957 4.85248 10.4082 5.12998 10.2582L6.38998 9.59821C6.66748 9.45571 6.77998 9.11071 6.62998 8.83321C6.48748 8.55571 6.14248 8.45071 5.86498 8.59321Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M6.77996 4.77569L5.93246 5.91569C5.74496 6.17069 5.79746 6.52319 6.05246 6.71069C6.30746 6.89819 6.65996 6.8457 6.84746 6.59069L7.69498 5.4507C7.88248 5.1957 7.82998 4.8432 7.57498 4.6557C7.32746 4.4682 6.96746 4.52069 6.77996 4.77569Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M11.1975 12.4257L12.045 11.2857C12.2325 11.0382 12.585 10.9857 12.84 11.1657C13.095 11.3532 13.1475 11.7057 12.96 11.9607L12.1125 13.1007C11.925 13.3557 11.5725 13.4082 11.3175 13.2207C11.0625 13.0332 11.01 12.6807 11.1975 12.4257Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M17.64 8.2557L18.5625 9.3357C18.765 9.5757 19.125 9.6057 19.365 9.4032C19.605 9.2007 19.635 8.8407 19.4325 8.6007L18.51 7.5207C18.3075 7.2807 17.9475 7.2507 17.7075 7.4532C17.4675 7.6632 17.4375 8.0157 17.64 8.2557Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M18.03 12.4107L17.3175 11.1807C17.16 10.9107 16.815 10.8132 16.5375 10.9707C16.2675 11.1282 16.17 11.4732 16.3275 11.7507L17.04 12.9807C17.1975 13.2507 17.5425 13.3482 17.82 13.1907C18.09 13.0332 18.1875 12.6882 18.03 12.4107Z",
+    fill: "#BEBEBE",
+  },
+  {
+    d: "M13.86 7.9107L15.165 8.4732C15.45 8.5932 15.585 8.9307 15.465 9.2232C15.3375 9.5157 15 9.6507 14.715 9.5232L13.41 8.9607C13.1175 8.8332 12.9825 8.4957 13.11 8.2107C13.2375 7.9182 13.575 7.7832 13.86 7.9107Z",
+    fill: "#BEBEBE",
+  },
+] as const;
+
+export const CREDIT_CARD_PATH_ITEMS = [
+  {
+    d: "M21.3143 18.5H3.68569C2.75669 18.5 2 17.7429 2 16.8134V6.6866C2 5.75705 2.75669 5 3.68569 5H21.3143C22.2433 5 23 5.75705 23 6.6866V16.8134C23 17.7505 22.2433 18.5 21.3143 18.5Z",
+    fill: "#00A6ED",
+  },
+  {
+    d: "M18.0398 14H4.71016C4.46122 14 4.25 13.7955 4.25 13.5379V11.4621C4.25 11.2121 4.45368 11 4.71016 11H18.0398C18.2888 11 18.5 11.2045 18.5 11.4621V13.5454C18.5 13.7955 18.2963 14 18.0398 14Z",
+    fill: "#F4F4F4",
+  },
+  {
+    d: "M19.5149 17H20.8601C21.2164 17 21.5 16.7164 21.5 16.3601V15.0149C21.5 14.6586 21.2164 14.375 20.8601 14.375H19.5149C19.1586 14.375 18.875 14.6586 18.875 15.0149V16.3601C18.875 16.7091 19.1659 17 19.5149 17Z",
+    fill: "#FFF478",
+  },
+  {
+    d: "M23 7.25H2V9.5H23V7.25Z",
+    fill: "#321B41",
+  },
+  {
+    d: "M5 12.5C5 12.2929 5.16789 12.125 5.375 12.125H13.625C13.8321 12.125 14 12.2929 14 12.5C14 12.7071 13.8321 12.875 13.625 12.875H5.375C5.16789 12.875 5 12.7071 5 12.5Z",
+    fill: "#9B9B9B",
+  },
+] as const;
+
+export const FULL_MOON_FACE_PATH_ITEMS = [
+  {
+    d: "M22.4993 11.9742C22.4993 17.773 17.7984 22.4738 11.9996 22.4738C6.20085 22.4738 1.5 17.773 1.5 11.9742C1.5 6.17539 6.20085 1.47455 11.9996 1.47455C17.7984 1.47455 22.4993 6.17539 22.4993 11.9742Z",
+    fill: "#FCD53F",
+  },
+  {
+    d: "M1.51645 11.3815C1.78137 11.2961 2.06393 11.25 2.35725 11.25C3.86963 11.25 5.09564 12.476 5.09564 13.9884C5.09564 15.4094 4.01326 16.5776 2.62797 16.7136C1.90653 15.2898 1.5 13.6795 1.5 11.9742C1.5 11.7753 1.50553 11.5777 1.51645 11.3815ZM19.6178 19.1995C18.358 20.5275 16.7546 21.5263 14.95 22.0536C14.7358 21.6633 14.614 21.2151 14.614 20.7384C14.614 19.226 15.84 18 17.3524 18C18.2943 18 19.1251 18.4755 19.6178 19.1995ZM22.4658 12.8182C22.3635 14.1043 22.0295 15.3251 21.5055 16.4389C20.6926 16.2346 20.0908 15.4989 20.0908 14.6227C20.0908 13.5884 20.9292 12.75 21.9635 12.75C22.1375 12.75 22.306 12.7737 22.4658 12.8182ZM11.6194 6C12.6536 6 13.492 5.16158 13.492 4.12733C13.492 3.09309 12.6536 2.25467 11.6194 2.25467C10.5851 2.25467 9.7467 3.09309 9.7467 4.12733C9.7467 5.16158 10.5851 6 11.6194 6ZM11.0921 16.9612C11.8799 17.578 12.0947 18.6196 11.5717 19.2875C11.0488 19.9554 9.98614 19.9968 9.19828 19.38C8.41042 18.7631 8.19567 17.7216 8.71863 17.0536C9.24159 16.3857 10.3042 16.3443 11.0921 16.9612Z",
+    fill: "#F9C23C",
+  },
+  {
+    d: "M8.00246 6.03003C6.97874 6.03003 6.14246 6.86046 6.14246 7.89003C6.14246 8.09714 5.97456 8.26503 5.76746 8.26503C5.56035 8.26503 5.39246 8.09714 5.39246 7.89003C5.39246 6.4446 6.56617 5.28003 8.00246 5.28003C8.20956 5.28003 8.37746 5.44792 8.37746 5.65503C8.37746 5.86214 8.20956 6.03003 8.00246 6.03003Z",
+    fill: "#321B41",
+  },
+  {
+    d: "M15.4125 5.65503C15.4125 5.44792 15.5804 5.28003 15.7875 5.28003C17.2238 5.28003 18.3975 6.4446 18.3975 7.89003C18.3975 8.09714 18.2296 8.26503 18.0225 8.26503C17.8154 8.26503 17.6475 8.09714 17.6475 7.89003C17.6475 6.86046 16.8112 6.03003 15.7875 6.03003C15.5804 6.03003 15.4125 5.86214 15.4125 5.65503Z",
+    fill: "#321B41",
+  },
+  {
+    d: "M7.73108 9.28799C7.88031 8.90772 8.25217 8.64001 8.69246 8.64001C9.14823 8.64001 9.53716 8.93555 9.6732 9.34851C9.77039 9.64357 10.0884 9.80397 10.3834 9.70677C10.6785 9.60958 10.8389 9.29159 10.7417 8.99652C10.4577 8.13448 9.64668 7.51501 8.69246 7.51501C7.78275 7.51501 6.99961 8.07231 6.68383 8.87704C6.57035 9.16623 6.71279 9.49266 7.00198 9.60614C7.29117 9.71962 7.6176 9.57718 7.73108 9.28799Z",
+    fill: "#321B41",
+  },
+  {
+    d: "M15.1575 8.64001C14.7172 8.64001 14.3453 8.90772 14.1961 9.28799C14.0826 9.57718 13.7562 9.71962 13.467 9.60614C13.1778 9.49266 13.0353 9.16623 13.1488 8.87704C13.4646 8.07231 14.2477 7.51501 15.1575 7.51501C16.1117 7.51501 16.9227 8.13448 17.2067 8.99652C17.3039 9.29159 17.1435 9.60958 16.8484 9.70677C16.5534 9.80397 16.2354 9.64357 16.1382 9.34851C16.0022 8.93555 15.6132 8.64001 15.1575 8.64001Z",
+    fill: "#321B41",
+  },
+  {
+    d: "M8.9777 12.4347C8.75803 12.2151 8.40188 12.2151 8.18221 12.4347C7.96254 12.6544 7.96254 13.0106 8.18221 13.2302C10.2394 15.2874 13.573 15.2874 15.6302 13.2302C15.8499 13.0106 15.8499 12.6544 15.6302 12.4347C15.4105 12.2151 15.0544 12.2151 14.8347 12.4347C13.2169 14.0526 10.5955 14.0526 8.9777 12.4347Z",
+    fill: "#321B41",
+  },
+] as const;
+
+export const LOUD_SPEAKER_PATH_ITEMS = [
+  {
+    d: "M18.7981 17.5535L17.9366 15.4485L16.6509 15.9747L17.5134 18.0822L17.5144 18.0847C17.6387 18.3852 17.4961 18.7159 17.207 18.8322L13.5276 20.3415L13.5257 20.3423C13.2252 20.4666 12.8945 20.324 12.7782 20.0349L11.9141 17.9235L10.6284 18.4497L11.4902 20.5555C11.8995 21.5696 13.0522 22.041 14.0558 21.6265L14.0569 21.626L17.7279 20.1201C18.7417 19.7108 19.213 18.5583 18.7987 17.555L18.7981 17.5535ZM6 17.625L7.28118 20.0842C7.74618 20.5717 7.72368 21.3442 7.22868 21.8017C6.74118 22.2592 5.96868 22.2367 5.51118 21.7492L1.82868 17.8417C1.37118 17.3542 1.39368 16.5817 1.88118 16.1242C2.36868 15.6667 3.14118 15.6892 3.59868 16.1767L6 17.625Z",
+    fill: "#FFB02E",
+  },
+  {
+    d: "M21.9213 11.3091L13.2588 2.12163C12.3063 1.10913 10.6338 1.37165 10.0413 2.63165L3.59875 16.1766L7.32626 20.1291L21.2313 14.4966C22.5138 13.9791 22.8738 12.3216 21.9213 11.3091Z",
+    fill: "#F9C23C",
+  },
+] as const;
+
+export const SCHOOL_BUS_PATH_ITEMS = [
+  {
+    d: "M22.5 7.455V15L12 16.5L1.5 15V7.455C1.5 6.6525 2.1525 6 2.955 6H21.045C21.855 6 22.5 6.6525 22.5 7.455Z",
+    fill: "#E6E6E6",
+  },
+  {
+    d: "M4.5 13.605V8.14501C4.5 7.79251 4.185 7.5 3.8025 7.5H1.5V14.25H3.8025C4.185 14.25 4.5 13.9575 4.5 13.605ZM6.6975 7.5H9.81C10.1925 7.5 10.5 7.79251 10.5075 8.14501V13.5975C10.5075 13.95 10.1925 14.2425 9.81 14.2425H6.6975C6.315 14.2425 6 13.9575 6 13.5975V8.14501C6 7.79251 6.315 7.5 6.6975 7.5ZM15.06 7.5H11.9475C11.565 7.5 11.25 7.79251 11.25 8.14501V13.5975C11.25 13.9575 11.565 14.2425 11.9475 14.2425H15.06C15.4425 14.2425 15.7575 13.95 15.7575 13.5975V8.14501C15.75 7.79251 15.4425 7.5 15.06 7.5ZM17.1975 7.5H20.31C20.6925 7.5 21 7.79251 21.0075 8.14501V13.5975C21.0075 13.95 20.6925 14.2425 20.31 14.2425H17.1975C16.815 14.2425 16.5 13.9575 16.5 13.5975V8.14501C16.5 7.79251 16.815 7.5 17.1975 7.5Z",
+    fill: "#26C9FC",
+  },
+  {
+    d: "M22.5 15V18.795C22.5 19.5975 21.8475 20.25 21.045 20.25H2.955C2.1525 20.25 1.5 19.5975 1.5 18.795C1.5 17.6916 1.5 16.2705 1.5 15H22.5Z",
+    fill: "#FFB02E",
+  },
+  {
+    d: "M1.5 7.5H3.8025C4.185 7.5 4.5 7.79251 4.5 8.14501V8.25V9H1.5V8.25V7.5ZM4.89 12.75H4.5975C4.1325 12.75 3.75 12.375 3.75 11.9025V10.5975C3.75 10.1325 4.1325 9.75 4.5975 9.75H4.89C5.085 9.75 5.25 9.90751 5.25 10.11V12.39C5.25 12.585 5.0925 12.75 4.89 12.75ZM9.0225 19.875C9.0225 21.3247 7.84725 22.5 6.3975 22.5C4.94775 22.5 3.7725 21.3247 3.7725 19.875C3.7725 18.4253 4.94775 17.25 6.3975 17.25C7.84725 17.25 9.0225 18.4253 9.0225 19.875ZM17.6325 22.5C19.0822 22.5 20.2575 21.3247 20.2575 19.875C20.2575 18.4253 19.0822 17.25 17.6325 17.25C16.1827 17.25 15.0075 18.4253 15.0075 19.875C15.0075 21.3247 16.1827 22.5 17.6325 22.5Z",
+    fill: "#212121",
+  },
+  {
+    d: "M22.5 15.75H22.125C21.4875 15.75 20.97 16.2825 21 16.9275C21.03 17.535 21.555 18 22.1625 18H22.5V15.75Z",
+    fill: "#F92F60",
+  },
+  {
+    d: "M2.1825 18.75H1.5V15.75H2.1825C2.6325 15.75 3 16.1175 3 16.5675V17.9325C3 18.3825 2.64 18.75 2.1825 18.75Z",
+    fill: "#FFF478",
+  },
+  {
+    d: "M7.5 19.875C7.5 20.4963 6.99632 21 6.375 21C5.75368 21 5.25 20.4963 5.25 19.875C5.25 19.2537 5.75368 18.75 6.375 18.75C6.99632 18.75 7.5 19.2537 7.5 19.875ZM18.75 19.875C18.75 20.4963 18.2463 21 17.625 21C17.0037 21 16.5 20.4963 16.5 19.875C16.5 19.2537 17.0037 18.75 17.625 18.75C18.2463 18.75 18.75 19.2537 18.75 19.875Z",
+    fill: "#D3D3D3",
+  },
+] as const;
+
+export const TENT_PATH_ITEMS = [
+  {
+    d: "M21.4569 10.0566C21.8338 9.52612 21.4534 9.1643 20.8029 9.18272C20.7705 9.18363 20.7381 9.18409 20.7055 9.18409C18.8416 9.18409 17.3586 7.67432 17.3586 5.81036C17.3586 5.6398 17.4462 5.2678 17.4706 5.10403C17.5666 4.46055 17.165 4.05008 16.5932 4.36045C15.3067 5.05877 14.1135 6.3143 14.1135 7.88104C14.1135 10.1592 15.9774 11.8467 18.2556 11.8467C19.6445 11.8467 20.7094 11.1087 21.4569 10.0566Z",
+    fill: "#FFB02E",
+  },
+  {
+    d: "M2.20841 22.5H6.8069L9.56591 21.637L12.3249 22.5C12.9697 22.5 13.5589 22.108 13.7924 21.52L17.8612 11.1642C17.9835 10.8593 17.75 10.5327 17.4165 10.5327H6.21054C6.01043 10.5327 5.83256 10.6525 5.76586 10.8376L1.55249 21.5744C1.3635 22.01 1.70814 22.5 2.20841 22.5Z",
+    fill: "#44911B",
+  },
+  {
+    d: "M13.7924 21.5091L17.8612 11.1534C17.9835 10.8485 17.7501 10.5218 17.4166 10.5218H8.08935C7.88924 10.5218 7.71138 10.6416 7.64468 10.8267L3.06444 22.5H6.81457L9.11104 21.928L12.3361 22.5C12.9697 22.5 13.5589 22.108 13.7924 21.5091Z",
+    fill: "#86D72F",
+  },
+  {
+    d: "M6.82204 22.4999H12.325C12.3694 22.4999 12.4139 22.4781 12.4584 22.4781V12.362H6.81093V22.4999H6.82204Z",
+    fill: "#1C1C1C",
+  },
+  {
+    d: "M5.92154 10.6198L10.2016 21.5091C10.4351 22.108 11.0131 22.4891 11.6691 22.4891H21.7856C22.2858 22.4891 22.6305 21.9991 22.4526 21.5418L18.2392 10.8049C18.1725 10.6198 17.9947 10.5 17.7945 10.5H17.4277H6.2217C6.09942 10.5218 5.99936 10.5653 5.92154 10.6198Z",
+    fill: "#86D72F",
+  },
+  {
+    d: "M15.7379 16.9683H13.5033C13.2143 16.9683 12.9586 16.7941 12.8474 16.5328L11.9136 14.2351C11.7246 13.7778 12.0692 13.2878 12.5695 13.2878H14.804C15.0931 13.2878 15.3488 13.462 15.4599 13.7233L16.3938 16.021C16.5828 16.4783 16.2381 16.9683 15.7379 16.9683Z",
+    fill: "#44911B",
+  },
+] as const;
