+++ packages/assets/src/icons/makehuman/index.ts
@@ -1,0 +1,9 @@
+export * from "./BarChart";
+export * from "./ConvenienceStore";
+export * from "./CookedRice";
+export * from "./CreditCard";
+export * from "./FullMoonFace";
+export * from "./LoudSpeaker";
+export * from "./MusicalNote";
+export * from "./SchoolBus";
+export * from "./Tent";
